refactor(cli): extract replay and source routes
